[core] `device_create_texture_view` -> `Texture::create_view`
