+++ wgpu-core/src/device/resource.rs
@@ -17,8 +17,7 @@ use arrayvec::ArrayVec;
 use bitflags::Flags;
 use smallvec::SmallVec;
 use wgt::{
-    math::align_to, DeviceLostReason, TextureFormat, TextureSampleType, TextureSelector,
-    TextureViewDimension,
+    math::align_to, DeviceLostReason, TextureFormat, TextureSampleType, TextureViewDimension,
 };
 
 #[cfg(feature = "trace")]
@@ -48,7 +47,7 @@ use crate::{
     resource::{
         self, Buffer, BufferState, ExternalTexture, ExternalTextureState, Labeled, ParentDevice,
         QuerySet, QuerySetState, RawResourceAccess, ResourceState, Sampler, StagingBuffer, Texture,
-        TextureView, TextureViewNotRenderableReason, TextureViewState, Tlas, TrackingData,
+        TextureView, Tlas, TrackingData,
     },
     resource_log,
     snatch::{SnatchGuard, SnatchLock, Snatchable},
@@ -2008,406 +2007,6 @@ impl Device {
             ));
         }
         texture
-    }
-
-    fn create_texture_view_inner(
-        self: &Arc<Self>,
-        texture: &Arc<Texture>,
-        desc: &resource::TextureViewDescriptor,
-    ) -> Result<Arc<TextureView>, resource::CreateTextureViewError> {
-        self.check_is_valid()?;
-
-        let snatch_guard = texture.device.snatchable_lock.read();
-
-        let texture_raw = texture.try_inner(&snatch_guard)?.raw();
-
-        // resolve TextureViewDescriptor defaults
-        // https://gpuweb.github.io/gpuweb/#abstract-opdef-resolving-gputextureviewdescriptor-defaults
-        let resolved_format = desc.format.unwrap_or_else(|| {
-            texture
-                .desc
-                .format
-                .aspect_specific_format(desc.range.aspect)
-                .unwrap_or(texture.desc.format)
-        });
-
-        let resolved_dimension = desc
-            .dimension
-            .unwrap_or_else(|| match texture.desc.dimension {
-                wgt::TextureDimension::D1 => TextureViewDimension::D1,
-                wgt::TextureDimension::D2 => {
-                    if texture.desc.array_layer_count() == 1 {
-                        TextureViewDimension::D2
-                    } else {
-                        TextureViewDimension::D2Array
-                    }
-                }
-                wgt::TextureDimension::D3 => TextureViewDimension::D3,
-            });
-
-        let resolved_mip_level_count = desc.range.mip_level_count.unwrap_or_else(|| {
-            texture
-                .desc
-                .mip_level_count
-                .saturating_sub(desc.range.base_mip_level)
-        });
-
-        let resolved_array_layer_count =
-            desc.range
-                .array_layer_count
-                .unwrap_or_else(|| match resolved_dimension {
-                    TextureViewDimension::D1
-                    | TextureViewDimension::D2
-                    | TextureViewDimension::D3 => 1,
-                    TextureViewDimension::Cube => 6,
-                    TextureViewDimension::D2Array | TextureViewDimension::CubeArray => texture
-                        .desc
-                        .array_layer_count()
-                        .saturating_sub(desc.range.base_array_layer),
-                });
-
-        let resolved_usage = {
-            let usage = desc.usage.unwrap_or(wgt::TextureUsages::empty());
-            if usage.is_empty() {
-                texture.desc.usage
-            } else if texture.desc.usage.contains(usage) {
-                // Transient texture usage subsetting is disallowed
-                if texture
-                    .desc
-                    .usage
-                    .contains(wgt::TextureUsages::TRANSIENT_ATTACHMENT)
-                    && texture.desc.usage != usage
-                {
-                    return Err(
-                        resource::CreateTextureViewError::InvalidTransientTextureViewUsage {
-                            texture: texture.desc.usage,
-                            view: usage,
-                        },
-                    );
-                }
-
-                usage
-            } else {
-                return Err(resource::CreateTextureViewError::InvalidTextureViewUsage {
-                    view: usage,
-                    texture: texture.desc.usage,
-                });
-            }
-        };
-
-        let format_features = self.describe_format_features(resolved_format)?;
-        let allowed_format_usages = format_features.allowed_usages;
-        if resolved_usage.contains(wgt::TextureUsages::RENDER_ATTACHMENT)
-            && !allowed_format_usages.contains(wgt::TextureUsages::RENDER_ATTACHMENT)
-        {
-            return Err(
-                resource::CreateTextureViewError::TextureViewFormatNotRenderable(resolved_format),
-            );
-        }
-
-        if resolved_usage.contains(wgt::TextureUsages::STORAGE_BINDING)
-            && !allowed_format_usages.contains(wgt::TextureUsages::STORAGE_BINDING)
-        {
-            return Err(
-                resource::CreateTextureViewError::TextureViewFormatNotStorage(resolved_format),
-            );
-        }
-
-        // validate TextureViewDescriptor
-
-        let aspects = hal::FormatAspects::new(texture.desc.format, desc.range.aspect);
-        if aspects.is_empty() {
-            return Err(resource::CreateTextureViewError::InvalidAspect {
-                texture_format: texture.desc.format,
-                requested_aspect: desc.range.aspect,
-            });
-        }
-
-        if desc.range.aspect == wgt::TextureAspect::All && resolved_format.is_multi_planar_format()
-        {
-            return Err(resource::CreateTextureViewError::MultiplanarFullTexture(
-                resolved_format,
-            ));
-        }
-
-        let format_is_good = if desc.range.aspect == wgt::TextureAspect::All {
-            resolved_format == texture.desc.format
-                || texture.desc.view_formats.contains(&resolved_format)
-        } else {
-            Some(resolved_format)
-                == texture
-                    .desc
-                    .format
-                    .aspect_specific_format(desc.range.aspect)
-        };
-        if !format_is_good {
-            return Err(resource::CreateTextureViewError::FormatReinterpretation {
-                texture: texture.desc.format,
-                view: resolved_format,
-            });
-        }
-
-        // check if multisampled texture is seen as anything but 2D
-        if texture.desc.sample_count > 1 && resolved_dimension != TextureViewDimension::D2 {
-            // Multisample is allowed on 2D arrays, only if explicitly supported
-            let multisample_array_exception = resolved_dimension == TextureViewDimension::D2Array
-                && self.features.contains(wgt::Features::MULTISAMPLE_ARRAY);
-
-            if !multisample_array_exception {
-                return Err(
-                    resource::CreateTextureViewError::InvalidMultisampledTextureViewDimension(
-                        resolved_dimension,
-                    ),
-                );
-            }
-        }
-
-        // check if the dimension is compatible with the texture
-        if texture.desc.dimension != resolved_dimension.compatible_texture_dimension() {
-            return Err(
-                resource::CreateTextureViewError::InvalidTextureViewDimension {
-                    view: resolved_dimension,
-                    texture: texture.desc.dimension,
-                },
-            );
-        }
-
-        match resolved_dimension {
-            TextureViewDimension::D1 | TextureViewDimension::D2 | TextureViewDimension::D3 => {
-                if resolved_array_layer_count != 1 {
-                    return Err(resource::CreateTextureViewError::InvalidArrayLayerCount {
-                        requested: resolved_array_layer_count,
-                        dim: resolved_dimension,
-                    });
-                }
-            }
-            TextureViewDimension::Cube => {
-                if resolved_array_layer_count != 6 {
-                    return Err(
-                        resource::CreateTextureViewError::InvalidCubemapTextureDepth {
-                            depth: resolved_array_layer_count,
-                        },
-                    );
-                }
-            }
-            TextureViewDimension::CubeArray => {
-                if !resolved_array_layer_count.is_multiple_of(6) {
-                    return Err(
-                        resource::CreateTextureViewError::InvalidCubemapArrayTextureDepth {
-                            depth: resolved_array_layer_count,
-                        },
-                    );
-                }
-            }
-            _ => {}
-        }
-
-        match resolved_dimension {
-            TextureViewDimension::Cube | TextureViewDimension::CubeArray => {
-                if texture.desc.size.width != texture.desc.size.height {
-                    return Err(resource::CreateTextureViewError::InvalidCubeTextureViewSize);
-                }
-            }
-            _ => {}
-        }
-
-        if resolved_mip_level_count == 0 {
-            return Err(resource::CreateTextureViewError::ZeroMipLevelCount);
-        }
-
-        let mip_level_end = desc
-            .range
-            .base_mip_level
-            .saturating_add(resolved_mip_level_count);
-
-        let level_end = texture.desc.mip_level_count;
-        if mip_level_end > level_end {
-            return Err(resource::CreateTextureViewError::TooManyMipLevels {
-                base_mip_level: desc.range.base_mip_level,
-                mip_level_count: resolved_mip_level_count,
-                total: level_end,
-            });
-        }
-
-        if resolved_array_layer_count == 0 {
-            return Err(resource::CreateTextureViewError::ZeroArrayLayerCount);
-        }
-
-        let array_layer_end = desc
-            .range
-            .base_array_layer
-            .saturating_add(resolved_array_layer_count);
-
-        let layer_end = texture.desc.array_layer_count();
-        if array_layer_end > layer_end {
-            return Err(resource::CreateTextureViewError::TooManyArrayLayers {
-                base_array_layer: desc.range.base_array_layer,
-                array_layer_count: resolved_array_layer_count,
-                total: layer_end,
-            });
-        };
-
-        // https://gpuweb.github.io/gpuweb/#abstract-opdef-renderable-texture-view
-        let render_extent = 'error: {
-            if !resolved_usage.contains(wgt::TextureUsages::RENDER_ATTACHMENT) {
-                break 'error Err(TextureViewNotRenderableReason::Usage(resolved_usage));
-            }
-
-            let allowed_view_dimensions = [
-                TextureViewDimension::D2,
-                TextureViewDimension::D2Array,
-                TextureViewDimension::D3,
-            ];
-            if !allowed_view_dimensions.contains(&resolved_dimension) {
-                break 'error Err(TextureViewNotRenderableReason::Dimension(
-                    resolved_dimension,
-                ));
-            }
-
-            if resolved_mip_level_count != 1 {
-                break 'error Err(TextureViewNotRenderableReason::MipLevelCount(
-                    resolved_mip_level_count,
-                ));
-            }
-
-            if resolved_array_layer_count != 1
-                && !(self.features.contains(wgt::Features::MULTIVIEW))
-            {
-                break 'error Err(TextureViewNotRenderableReason::ArrayLayerCount(
-                    resolved_array_layer_count,
-                ));
-            }
-
-            if !texture.desc.format.is_multi_planar_format()
-                && aspects != hal::FormatAspects::from(texture.desc.format)
-            {
-                break 'error Err(TextureViewNotRenderableReason::Aspects(aspects));
-            }
-
-            Ok(texture
-                .desc
-                .compute_render_extent(desc.range.base_mip_level, desc.range.aspect.to_plane()))
-        };
-
-        // filter the usages based on the other criteria
-        let usage = {
-            let resolved_hal_usage = conv::map_texture_usage(
-                resolved_usage,
-                resolved_format.into(),
-                format_features.flags,
-            );
-            let mask_copy = !(wgt::TextureUses::COPY_SRC | wgt::TextureUses::COPY_DST);
-            let mask_dimension = match resolved_dimension {
-                TextureViewDimension::Cube | TextureViewDimension::CubeArray => {
-                    wgt::TextureUses::RESOURCE
-                }
-                TextureViewDimension::D3 => {
-                    wgt::TextureUses::RESOURCE
-                        | wgt::TextureUses::STORAGE_READ_ONLY
-                        | wgt::TextureUses::STORAGE_WRITE_ONLY
-                        | wgt::TextureUses::STORAGE_READ_WRITE
-                }
-                _ => wgt::TextureUses::all(),
-            };
-            let mask_mip_level = if resolved_mip_level_count == 1 {
-                wgt::TextureUses::all()
-            } else {
-                wgt::TextureUses::RESOURCE
-            };
-            resolved_hal_usage & mask_copy & mask_dimension & mask_mip_level
-        };
-
-        // use the combined depth-stencil format for the view
-        let format = if resolved_format.is_depth_stencil_component(texture.desc.format) {
-            texture.desc.format
-        } else {
-            resolved_format
-        };
-
-        let resolved_range = wgt::ImageSubresourceRange {
-            aspect: desc.range.aspect,
-            base_mip_level: desc.range.base_mip_level,
-            mip_level_count: Some(resolved_mip_level_count),
-            base_array_layer: desc.range.base_array_layer,
-            array_layer_count: Some(resolved_array_layer_count),
-        };
-
-        let hal_desc = hal::TextureViewDescriptor {
-            label: desc.label.to_hal(self.instance_flags),
-            format,
-            dimension: resolved_dimension,
-            usage,
-            range: resolved_range,
-        };
-
-        let raw = unsafe { self.raw().create_texture_view(texture_raw, &hal_desc) }
-            .map_err(|e| self.handle_hal_error(e))?;
-
-        let selector = TextureSelector {
-            mips: desc.range.base_mip_level..mip_level_end,
-            layers: desc.range.base_array_layer..array_layer_end,
-        };
-
-        let view = TextureView {
-            state: ResourceState::Valid(TextureViewState {
-                raw: Snatchable::new(raw),
-                render_extent,
-            }),
-            parent: texture.clone(),
-            device: self.clone(),
-            desc: resource::HalTextureViewDescriptor {
-                texture_format: texture.desc.format,
-                format: resolved_format,
-                dimension: resolved_dimension,
-                usage: resolved_usage,
-                range: resolved_range,
-            },
-            format_features: texture.format_features,
-            samples: texture.desc.sample_count,
-            selector,
-            label: desc.label.to_string(),
-        };
-
-        let view = Arc::new(view);
-
-        {
-            let mut views = texture.views.lock();
-            views.push(Arc::downgrade(&view));
-        }
-
-        Ok(view)
-    }
-
-    pub fn create_texture_view(
-        self: &Arc<Self>,
-        texture: &Arc<Texture>,
-        desc: &resource::TextureViewDescriptor,
-    ) -> (Arc<TextureView>, Option<resource::CreateTextureViewError>) {
-        profiling::scope!("Texture::create_view");
-
-        let (view, error) = match self.create_texture_view_inner(texture, desc) {
-            Ok(view) => (view, None),
-            Err(e) => (TextureView::invalid(self, texture, desc), Some(e)),
-        };
-
-        api_log!(
-            "Texture::create_view({:?}) -> {:?}",
-            Arc::as_ptr(texture),
-            Arc::as_ptr(&view)
-        );
-
-        #[cfg(feature = "trace")]
-        if let Some(ref mut trace) = *self.trace.lock() {
-            use crate::device::trace;
-            use trace::IntoTrace as _;
-            trace.add(trace::Action::CreateTextureView {
-                id: view.to_trace(),
-                parent: texture.to_trace(),
-                desc: desc.clone(),
-            });
-        }
-
-        (view, error)
     }
 
     pub fn create_external_texture(
